Dagger hilt and retrofit implementation added

+++ app/src/main/java/com/multi/trivia/data/model/QuizResponse.kt
@@ -3,6 +3,6 @@ package com.multi.trivia.data.model
 import com.google.gson.annotations.SerializedName
 
 data class QuizResponse(
-    @SerializedName("response_code") val responseCode: Long,
+    @SerializedName("response_code") val responseCode: Int,
     @SerializedName("results") val questionList: List<Question>
 )
